added impl method send Mailing, some refactored    - added UtilsMail    - added impl method for MailSenderServiceImpl {int sendAnonymously(Long id, Message message)}    - refactored logic from send methods

+++ src/main/java/org/gnori/mailsenderbot/command/commands/SendAnonymouslyCommand.java
@@ -4,9 +4,12 @@ import org.gnori.mailsenderbot.command.Command;
 import org.gnori.mailsenderbot.entity.MessageSentRecord;
 import org.gnori.mailsenderbot.model.Message;
 import org.gnori.mailsenderbot.repository.MessageRepository;
+import org.gnori.mailsenderbot.service.MailSenderService;
 import org.gnori.mailsenderbot.service.ModifyDataBaseService;
 import org.gnori.mailsenderbot.service.SendBotMessageService;
 import org.telegram.telegrambots.meta.api.objects.Update;
+
+import java.util.Collections;
 
 import static org.gnori.mailsenderbot.command.commands.Utils.prepareCallbackDataForBeginningMessage;
 
@@ -14,32 +17,39 @@ public class SendAnonymouslyCommand implements Command {
     private final SendBotMessageService sendBotMessageService;
     private final ModifyDataBaseService modifyDataBaseService;
     private final MessageRepository messageRepository;
+    private final MailSenderService mailSenderService;
 
     public SendAnonymouslyCommand(SendBotMessageService sendBotMessageService,
                                   ModifyDataBaseService modifyDataBaseService,
-                                  MessageRepository messageRepository) {
+                                  MessageRepository messageRepository,
+                                  MailSenderService mailSenderService) {
         this.sendBotMessageService = sendBotMessageService;
         this.modifyDataBaseService = modifyDataBaseService;
         this.messageRepository = messageRepository;
+        this.mailSenderService = mailSenderService;
     }
 
     @Override
     public void execute(Update update) {
-        //TODO added impl service for send Mail
-        // and send Success alert for telegram
         var chatId =  update.getCallbackQuery().getMessage().getChatId();
         var messageId = update.getCallbackQuery().getMessage().getMessageId();
         var newCallbackData = prepareCallbackDataForBeginningMessage();
-        var text = "отправлено:✔"+
-                "\nВыберите необходимый пункт👇🏿";
         var messageToSend = messageRepository.getMessage(chatId);
-        //TODO send mail logic
 
-        createAndAddMessageSentRecord(chatId, messageToSend);
-        messageRepository.removeMessage(chatId);
+        sendBotMessageService.executeEditMessage(chatId,messageId,"Производится отправка...🛫", Collections.emptyList(),false);
+
+        var sendResult = mailSenderService.sendAnonymously(chatId,messageToSend);
+        var text = "неотправлено:❌";
+        if (sendResult==1){
+            text="отправлено✔";
+            createAndAddMessageSentRecord(chatId, messageToSend);
+            messageRepository.removeMessage(chatId);
+        }
+        text += "\nВыберите необходимый пункт👇🏿";
+
         sendBotMessageService.executeEditMessage(chatId,messageId,text,newCallbackData,false);
-
     }
+
     private void createAndAddMessageSentRecord(Long id, Message message){
         var countMessages = message.getRecipients().size() * message.getCountForRecipient();
         var messageSentRecord = MessageSentRecord.builder().countMessages(countMessages).build();
